# specs/001-trade-show-intelligence/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Trade Show Intelligence Platform API
  description: |
    REST API for badge scan upload, enrichment processing, persona management, and report generation.

    **Key Features**:
    - CSV badge scan upload and processing
    - Batch enrichment with multi-LLM verification
    - Persona-based lead scoring and tier assignment
    - Report generation with filtering and export
    - Storage adapter configuration (Local, MySQL, HubSpot)
  version: 1.0.0
  contact:
    name: Trade Show Intelligence Platform
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://app.tradeshowai.com/api
    description: Production server
...
paths:
...
  /personas:
    get:
      summary: List all personas
      description: Retrieve all persona templates
      tags:
        - Personas
      parameters:
        - name: defaultOnly
          in: query
          required: false
          schema:
            type: boolean
          description: Return only default personas
      responses:
        '200':
          description: Personas retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - personas
                properties:
                  personas:
                    type: array
                    items:
                      $ref: '#/components/schemas/Persona'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: Create new persona
      description: Create custom persona template
      tags:
        - Personas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaInput'
      responses:
        '201':
          description: Persona created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Persona'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    Persona:
      type: object
      required:
        - id
        - name
        - isDefault
        - criteria
        - weights
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        isDefault:
          type: boolean
        criteria:
          $ref: '#/components/schemas/PersonaCriteria'
        weights:
          $ref: '#/components/schemas/PersonaWeights'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
    PersonaInput:
      type: object
      required:
        - name
        - criteria
        - weights
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        isDefault:
          type: boolean
        criteria:
          $ref: '#/components/schemas/PersonaCriteria'
        weights:
          $ref: '#/components/schemas/PersonaWeights'
    PersonaCriteria:
      type: object
      properties:
        companySizeRange:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
        industries:
          type: array
          items:
            type: string
        technologies:
          type: array
          items:
            type: string
        revenueRange:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
        geographies:
          type: array
          items:
            type: string
        decisionMakerTitles:
          type: array
          items:
            type: string
        fundingStages:
          type: array
          items:
            type: string
    PersonaWeights:
      type: object
      required:
        - companySize
        - industry
        - technology
        - revenue
        - geography
        - decisionMaker
        - fundingStage
      properties:
        companySize:
          type: number
          minimum: 0
          maximum: 1
        industry:
          type: number
          minimum: 0
          maximum: 1
        technology:
          type: number
          minimum: 0
          maximum: 1
        revenue:
          type: number
          minimum: 0
          maximum: 1
        geography:
          type: number
          minimum: 0
          maximum: 1
        decisionMaker:
          type: number
          minimum: 0
          maximum: 1
        fundingStage:
          type: number
          minimum: 0
          maximum: 1
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
security: []